Updates for version 25.2401.4023
Bug-fix only; A **Camera3D** can now be orbited around a subject, crossing look-at lines on  the horizontal and vertical planes without an orientation. The hack from the previous version has been removed and replaced with the logic to avoid a flipping condition. Please keep in mind this camera implements an "always up" style of view, relative to its local Y axis. When passing from one side of the look-at X or Z planes, the view will flip to keep the camera upright for that relationship.